address PR review feedback: quote logging, OpenAPI sync, page_size support, u16 types
- Log warning on quote fetch failure before falling back to "-" in orders list
- Add 422 response to OpenAPI annotation for get_orders_by_token
- Wire page_size parameter through trait, implementation, and route handler
- Change page and page_size params from u32 to u16 to match SDK types
- Update rain.orderbook submodule with page_size support in get_orders

+++ src/types/orders.rs
@@ -10,10 +10,10 @@ use utoipa::{IntoParams, ToSchema};
 pub struct OrdersPaginationParams {
     #[field(name = "page")]
     #[param(example = 1)]
-    pub page: Option<u32>,
+    pub page: Option<u16>,
     #[field(name = "pageSize")]
     #[param(example = 20)]
-    pub page_size: Option<u32>,
+    pub page_size: Option<u16>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromFormField, ToSchema)]
@@ -31,10 +31,10 @@ pub struct OrdersByTokenParams {
     pub side: Option<OrderSide>,
     #[field(name = "page")]
     #[param(example = 1)]
-    pub page: Option<u32>,
+    pub page: Option<u16>,
     #[field(name = "pageSize")]
     #[param(example = 20)]
-    pub page_size: Option<u32>,
+    pub page_size: Option<u16>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
